requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pytz>=2023.3
